importarProductos5

+++ src/db/models/producto.model.js
@@ -35,7 +35,7 @@ const ProductoSchema = {
   },
   impuesto:{
     allowNull: false,
-    type: DataTypes.INTEGER
+    type: DataTypes.DOUBLE
   },
   rubro: {
     allowNull: false,
